requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
lmdb>=1.4.0
wandb>=0.15.0
accelerate>=0.20.0
safetensors>=0.3.0
packaging>=21.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
